requirements.txt
alembic==1.3.2
astroid==2.3.3
blinker==1.4
Bootstrap-Flask==1.2.0
Click==7.0
colorama==0.4.3
dominate==2.4.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
flask-talisman==0.7.0
Flask-WTF==0.14.2
gunicorn==20.0.4
h11==0.8.1
httptools==0.0.13
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
uvicorn==0.10.8
uvloop==0.14.0
visitor==0.1.3
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
